feat: Merge common stuff of sandboxes; Refactor elements into multiple files.
